requirements.txt
fastapi>=0.115.0
python-binance>=1.0.19
python-dotenv>=1.0.1
streamlit>=1.44.0
uvicorn>=0.34.0
requests>=2.32.0

